requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.4.0
certifi==2024.6.2
cffi==2.0.0
charset-normalizer==3.3.2
cmake==4.2.1
comtypes==1.4.14
contourpy==1.1.0
cryptography==46.0.2
cycler==0.11.0
dateparser==1.2.2
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dlib-bin==20.0.0
filelock==3.19.1
fonttools==4.42.0
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.7
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.7.2
multidict==6.7.1
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==23.1
pandas==2.2.2
Pillow==10.0.0
pipenv==2025.0.4
platformdirs==4.3.8
propcache==0.4.1
psycopg2-binary==2.9.10
PyAudio==0.2.14
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pyparsing==3.0.9
pypiwin32==223
python-binance==1.0.34
python-dateutil==2.8.2
python-dotenv==1.2.1
python3-openid==3.2.0
pyttsx3==2.99
pytz==2025.2
pywin32==311
regex==2026.1.15
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.16.3
six==1.16.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
SpeechRecognition==3.10.4
sqlparse==0.5.3
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.3.1
urllib3==2.2.1
virtualenv==20.34.0
websockets==16.0
whitenoise==6.9.0
yarl==1.22.0
